requirements.txt
faiss-cpu
torch
numpy
pymupdf
langchain
langchain_huggingface
langchain_experimental
langchain_ollama
stix2
hnswlib
lancedb
datasets
ragas
chromadb
rapidfuzz
matplotlib
unstructured
nltk
evaluate
rouge_score
bert_score